+++ Fabric/src/main/java/fuzs/forgeconfigapiport/fabric/impl/config/ForgeConfigApiPortConfig.java
@@ -9,7 +9,6 @@ import com.electronwill.nightconfig.toml.TomlFormat;
 import com.google.common.collect.ImmutableMap;
 import fuzs.forgeconfigapiport.impl.ForgeConfigAPIPort;
 import net.fabricmc.loader.api.FabricLoader;
-import net.neoforged.fml.config.ModConfig;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -21,7 +20,11 @@ import java.util.Objects;
 public class ForgeConfigApiPortConfig {
     public static final ForgeConfigApiPortConfig INSTANCE;
     private static final String CONFIG_FILE_NAME = ForgeConfigAPIPort.MOD_ID + ".toml";
-    private static final Map<String, Object> CONFIG_VALUES = ImmutableMap.<String, Object>builder().put("defaultConfigsPath", "defaultconfigs").put("forceGlobalServerConfigs", true).put("recreateConfigsWhenParsingFails", true).put("correctConfigValuesFromDefaultConfig", true).put("disableConfigWatcher", false).build();
+    private static final Map<String, Object> CONFIG_VALUES = ImmutableMap.<String, Object>builder()
+            .put("defaultConfigsPath", "defaultconfigs")
+            .put("correctConfigValuesFromDefaultConfig", true)
+            .put("disableConfigWatcher", false)
+            .build();
     private static final ConfigSpec CONFIG_SPEC;
 
     static {
@@ -40,15 +43,31 @@ public class ForgeConfigApiPortConfig {
 
     // Copied from FML config
     private void loadFrom(final Path configFile) {
-        this.configData = CommentedFileConfig.builder(configFile, TomlFormat.instance()).sync().onFileNotFound(FileNotFoundAction.copyData(Objects.requireNonNull(this.getClass().getResourceAsStream("/" + CONFIG_FILE_NAME)))).autosave().autoreload().writingMode(WritingMode.REPLACE).build();
+        this.configData = CommentedFileConfig.builder(configFile, TomlFormat.instance())
+                .sync()
+                .onFileNotFound(FileNotFoundAction.copyData(Objects.requireNonNull(this.getClass()
+                        .getResourceAsStream("/" + CONFIG_FILE_NAME))))
+                .autosave()
+                .autoreload()
+                .writingMode(WritingMode.REPLACE)
+                .build();
         try {
             this.configData.load();
         } catch (ParsingException e) {
-            throw new RuntimeException("Failed to load %s config from %s".formatted(ForgeConfigAPIPort.MOD_NAME, configFile), e);
+            throw new RuntimeException("Failed to load %s config from %s".formatted(ForgeConfigAPIPort.MOD_NAME,
+                    configFile
+            ), e);
         }
         if (!CONFIG_SPEC.isCorrect(this.configData)) {
             ForgeConfigAPIPort.LOGGER.warn("Configuration file {} is not correct. Correcting", configFile);
-            CONFIG_SPEC.correct(this.configData, (action, path, incorrectValue, correctedValue) -> ForgeConfigAPIPort.LOGGER.warn("Incorrect key {} was corrected from {} to {}", path, incorrectValue, correctedValue));
+            CONFIG_SPEC.correct(this.configData,
+                    (action, path, incorrectValue, correctedValue) -> ForgeConfigAPIPort.LOGGER.warn(
+                            "Incorrect key {} was corrected from {} to {}",
+                            path,
+                            incorrectValue,
+                            correctedValue
+                    )
+            );
         }
         this.configData.save();
         getOrCreateGameRelativePath(Paths.get(this.<String>getValue("defaultConfigsPath")));
@@ -77,13 +96,5 @@ public class ForgeConfigApiPortConfig {
 
     public static Path getDefaultConfigsDirectory() {
         return FabricLoader.getInstance().getGameDir().resolve(INSTANCE.<String>getValue("defaultConfigsPath"));
-    }
-
-    public static Path getConfigPath(Path configBasePath, String fileName) {
-        Path configPath = configBasePath.resolve(fileName);
-        if (INSTANCE.<Boolean>getValue("forceGlobalServerConfigs") && Files.notExists(configPath)) {
-            configPath = FabricLoader.getInstance().getConfigDir().resolve(fileName);
-        }
-        return configPath;
     }
 }